automation-store/requirements/base.txt
Django==4.0.4
django-extensions==3.1.5
djangorestframework==3.13.1
python-decouple==3.6